add fuzz example

+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub fn add(a: i32, b: i32) -> i32 {
+pub fn add(a: u128, b: u128) -> u128 {
     a + b
 }
